requirements.txt
torch-scatter
torch-sparse
torch-cluster
torch-spline-conv
torch-geometric

pytorch-lightning

rdkit
numpy
networkx
pandas
matplotlib